[Java] Move mature tooling into aeron client package
