+++ aeron-client/src/main/java/io/aeron/reports/LossReportUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.aeron.driver.reports;
+package io.aeron.reports;
 
 import java.io.File;
 import java.nio.MappedByteBuffer;
