+++ aeron-client/src/main/java/io/aeron/status/CounterKeyOffset.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2014-2024 Real Logic Limited.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.aeron.status;
+
+import org.agrona.concurrent.status.CountersReader;
+
+import static org.agrona.BitUtil.SIZE_OF_INT;
+import static org.agrona.BitUtil.SIZE_OF_LONG;
+
+/**
+ * Offset locations within the Metadata Key of Stream Counters.
+ */
+public class CounterKeyOffset
+{
+    /**
+     * Offset in the key metadata for the registration id of the counter.
+     */
+    public static final int REGISTRATION_ID_OFFSET = 0;
+
+    /**
+     * Offset in the key metadata for the session id of the counter.
+     */
+    public static final int SESSION_ID_OFFSET = REGISTRATION_ID_OFFSET + SIZE_OF_LONG;
+
+    /**
+     * Offset in the key metadata for the stream id of the counter.
+     */
+    public static final int STREAM_ID_OFFSET = SESSION_ID_OFFSET + SIZE_OF_INT;
+
+    /**
+     * Offset in the key metadata for the channel of the counter.
+     */
+    public static final int CHANNEL_OFFSET = STREAM_ID_OFFSET + SIZE_OF_INT;
+
+    /**
+     * The maximum length in bytes of the encoded channel identity.
+     */
+    public static final int MAX_CHANNEL_LENGTH = CountersReader.MAX_KEY_LENGTH - (CHANNEL_OFFSET + SIZE_OF_INT);
+}
