+++ aeron-client/src/test/java/io/aeron/reports/LossReportTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.aeron.driver.reports;
+package io.aeron.reports;
 
 import org.agrona.BitUtil;
 import org.agrona.concurrent.AtomicBuffer;
@@ -23,7 +23,7 @@ import org.mockito.InOrder;
 
 import java.nio.ByteBuffer;
 
-import static io.aeron.driver.reports.LossReport.*;
+import static io.aeron.reports.LossReport.*;
 import static org.agrona.BitUtil.SIZE_OF_INT;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
