+++ aeron-client/src/test/java/io/aeron/tooling/LogInspectorAsciiFormatBytesTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.aeron.samples;
+package io.aeron.tooling;
 
 import org.agrona.concurrent.UnsafeBuffer;
 import org.junit.jupiter.api.AfterEach;
