+++ aeron-driver/src/main/java/io/aeron/driver/Configuration.java
@@ -233,13 +233,13 @@ public final class Configuration
     public static final int ERROR_BUFFER_LENGTH_DEFAULT = 1024 * 1024;
 
     /**
-     * Property name for length of the memory mapped buffer for the {@link io.aeron.driver.reports.LossReport}.
+     * Property name for length of the memory mapped buffer for the {@link io.aeron.reports.LossReport}.
      */
     @Config
     public static final String LOSS_REPORT_BUFFER_LENGTH_PROP_NAME = "aeron.loss.report.buffer.length";
 
     /**
-     * Default buffer length for the {@link io.aeron.driver.reports.LossReport}.
+     * Default buffer length for the {@link io.aeron.reports.LossReport}.
      */
     @Config
     public static final int LOSS_REPORT_BUFFER_LENGTH_DEFAULT = 1024 * 1024;
@@ -1754,9 +1754,9 @@ public final class Configuration
     }
 
     /**
-     * Length of the memory mapped buffer for the {@link io.aeron.driver.reports.LossReport}.
+     * Length of the memory mapped buffer for the {@link io.aeron.reports.LossReport}.
      *
-     * @return length of the memory mapped buffer for the {@link io.aeron.driver.reports.LossReport}.
+     * @return length of the memory mapped buffer for the {@link io.aeron.reports.LossReport}.
      * @see #LOSS_REPORT_BUFFER_LENGTH_PROP_NAME
      */
     public static int lossReportBufferLength()
