+++ aeron-driver/src/main/java/io/aeron/driver/MediaDriver.java
@@ -23,7 +23,7 @@ import io.aeron.driver.buffer.FileStoreLogFactory;
 import io.aeron.driver.buffer.LogFactory;
 import io.aeron.driver.exceptions.ActiveDriverException;
 import io.aeron.driver.media.*;
-import io.aeron.driver.reports.LossReport;
+import io.aeron.reports.LossReport;
 import io.aeron.driver.status.DutyCycleStallTracker;
 import io.aeron.driver.status.SystemCounters;
 import io.aeron.exceptions.AeronException;
@@ -58,10 +58,10 @@ import java.util.function.Consumer;
 
 import static io.aeron.CncFileDescriptor.*;
 import static io.aeron.driver.Configuration.*;
-import static io.aeron.driver.reports.LossReportUtil.mapLossReport;
 import static io.aeron.driver.status.SystemCounterDescriptor.CONTROLLABLE_IDLE_STRATEGY;
 import static io.aeron.driver.status.SystemCounterDescriptor.*;
 import static io.aeron.logbuffer.LogBufferDescriptor.TERM_MAX_LENGTH;
+import static io.aeron.reports.LossReportUtil.mapLossReport;
 import static java.nio.charset.StandardCharsets.US_ASCII;
 import static java.util.concurrent.atomic.AtomicIntegerFieldUpdater.newUpdater;
 import static org.agrona.BitUtil.SIZE_OF_LONG;
