+++ aeron-driver/src/main/java/io/aeron/driver/PublicationImage.java
@@ -20,13 +20,13 @@ import io.aeron.driver.buffer.RawLog;
 import io.aeron.driver.media.ImageConnection;
 import io.aeron.driver.media.ReceiveChannelEndpoint;
 import io.aeron.driver.media.ReceiveDestinationTransport;
-import io.aeron.driver.reports.LossReport;
 import io.aeron.driver.status.SystemCounters;
 import io.aeron.logbuffer.LogBufferDescriptor;
 import io.aeron.logbuffer.TermRebuilder;
 import io.aeron.protocol.DataHeaderFlyweight;
 import io.aeron.protocol.RttMeasurementFlyweight;
 import io.aeron.protocol.StatusMessageFlyweight;
+import io.aeron.reports.LossReport;
 import org.agrona.CloseHelper;
 import org.agrona.ErrorHandler;
 import org.agrona.SystemUtil;
