+++ aeron-driver/src/main/java/io/aeron/driver/status/ReceiverHwm.java
@@ -16,6 +16,7 @@
 package io.aeron.driver.status;
 
 import io.aeron.AeronCounters;
+import io.aeron.status.CounterLabelNameDescriptor;
 import org.agrona.MutableDirectBuffer;
 import org.agrona.concurrent.status.CountersManager;
 import org.agrona.concurrent.status.UnsafeBufferPosition;
@@ -34,7 +35,7 @@ public class ReceiverHwm
     /**
      * Human-readable name for the counter.
      */
-    public static final String NAME = "rcv-hwm";
+    public static final String NAME = CounterLabelNameDescriptor.RECEIVER_HWM;
 
     /**
      * Allocate a new receiver high-water-mark position counter for a stream.
